Fix build issue

+++ src/runtime_src/core/pcie/driver/linux/include/mailbox_proto.h
@@ -154,12 +154,12 @@ struct xcl_sensor {
 	uint32_t config_mode;
 	uint32_t hbm_temp1;
 	uint32_t hbm_temp2;
-	uint64_t cur_3v3_pex;
-	uint64_t cur_0v85;
-	uint64_t vol_3v3_vcc;
-	uint64_t vol_1v2_hbm;
-	uint64_t vol_2v5_vpp;
-	uint64_t vccint_bram;
+	uint32_t cur_3v3_pex;
+	uint32_t cur_0v85;
+	uint32_t vol_3v3_vcc;
+	uint32_t vol_1v2_hbm;
+	uint32_t vol_2v5_vpp;
+	uint32_t vccint_bram;
 };
 
 /**
